Encapsulating strVectorJoin to utilities file
