+++ src/utils.h
@@ -1,0 +1,9 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+#include <string>
+#include <vector>
+
+std::string strVectorJoin(const std::vector<std::string> &elements, const std::string &separator);
+
+#endif
